Modularize code and add comments
