+++ Jarvis/jarvis.py
@@ -1,3 +1,4 @@
+#import dependencies
 import pyttsx3
 import pywin32_system32
 import datetime
@@ -8,124 +9,82 @@ import os
 import random
 import pyautogui
 
-engine = pyttsx3.init()
+#import javis actions
+from jarvis_actions import speak
+from jarvis_actions import time
+from jarvis_actions import date
+from jarvis_actions import wishme
+from jarvis_actions import screenshot
+from jarvis_actions import takecommand
 
 
-def speak(audio):
-    engine.say(audio)
-    engine.runAndWait()
-
-
-def time():
-    Time = datetime.datetime.now().strftime("%I:%M:%S")
-    speak("the current time is")
-    speak(Time)
-    print("The current time is ", Time)
-
-
-def date():
-    day = int(datetime.datetime.now().day)
-    month = int(datetime.datetime.now().month)
-    year = int(datetime.datetime.now().year)
-    speak("the current date is")
-    speak(day)
-    speak(month)
-    speak(year)
-    print("The current date is " + str(day) + "/" + str(month) + "/" + str(year))
-
-
-def wishme():
-    print("Welcome back sir!!")
-    speak("Welcome back sir!!")
-
-    hour = datetime.datetime.now().hour
-    if 4 <= hour < 12:
-        speak("Good Morning Sir!!")
-        print("Good Morning Sir!!")
-    elif 12 <= hour < 16:
-        speak("Good Afternoon Sir!!")
-        print("Good Afternoon Sir!!")
-    elif 16 <= hour < 24:
-        speak("Good Evening Sir!!")
-        print("Good Evening Sir!!")
-    else:
-        speak("Good Night Sir, See You Tommorrow")
-
-    speak("Jarvis at your service sir, please tell me how may I help you.")
-    print("Jarvis at your service sir, please tell me how may I help you.")
-
-
-def screenshot():
-    img = pyautogui.screenshot()
-    img_path = os.path.expanduser("~\\Pictures\\ss.png")
-    img.save(img_path)
-
-
-def takecommand():
-    r = sr.Recognizer()
-    with sr.Microphone() as source:
-        print("Listening...")
-        r.pause_threshold = 1
-        audio = r.listen(source)
-
-    try:
-        print("Recognizing...")
-        query = r.recognize_google(audio, language="en-in")
-        print(query)
-
-    except Exception as e:
-        print(e)
-        speak("Please say that again")
-        return "Try Again"
-
-    return query
-
-
+#start program
 if __name__ == "__main__":
+    #greet user
     wishme()
+
     while True:
+        #acquire user input
         query = takecommand().lower()
+
+        #if the user mentions the time, report what time it is
         if "time" in query:
             time()
 
+        #if the user mentions the date, report what the date is
         elif "date" in query:
             date()
 
+        #if the user asks for information about the personal assistant, it is provided
         elif "who are you" in query:
             speak("I'm JARVIS created by Mr. Kishan and I'm a desktop voice assistant.")
             print("I'm JARVIS created by Mr. Kishan and I'm a desktop voice assistant.")
 
+        #if the user asks the assistant how they are, an appropriate response is given
         elif "how are you" in query:
             speak("I'm fine sir, What about you?")
             print("I'm fine sir, What about you?")
 
+        #if the user says that they are fine, then an appropriate response is given
         elif "fine" in query:
             speak("Glad to hear that sir!!")
             print("Glad to hear that sir!!")
 
+        #if the user says that they are good, an appropriate response is given
         elif "good" in query:
             speak("Glad to hear that sir!!")
             print("Glad to hear that sir!!")
 
+        #if the user wishes to recieve a wikipedia summary
         elif "wikipedia" in query:
             try:
+                #feedback to let the user know that the assistant is working on it
                 speak("Ok wait sir, I'm searching...")
+
+                #remove the word wikipedia from the query and produce a result
                 query = query.replace("wikipedia", "")
                 result = wikipedia.summary(query, sentences=2)
+
+                #output result
                 print(result)
                 speak(result)
             except:
+                #if unsuccessful, inform the user
                 speak("Can't find this page sir, please ask something else")
 
+        #open youtube.com in default browser
         elif "open youtube" in query:
             wb.open("youtube.com")
 
+        #open google.com in default browser
         elif "open google" in query:
             wb.open("google.com")
 
+        #open stack overflow in default browser
         elif "open stack overflow" in query:
             wb.open("stackoverflow.com")
 
+        #play music from a playlist
         elif "play music" in query:
             song_dir = os.path.expanduser("~\\Music")
             songs = os.listdir(song_dir)
@@ -134,10 +93,12 @@ if __name__ == "__main__":
             y = random.randint(0, x)
             os.startfile(os.path.join(song_dir, songs[y]))
 
+        #open the chrome app
         elif "open chrome" in query:
             chromePath = "C:\\Program Files (x86)\\Google\\Chrome\\Application\\chrome.exe"
             os.startfile(chromePath)
 
+        #open a chrome tab with a google search
         elif "search on chrome" in query:
             try:
                 speak("What should I search?")
@@ -147,11 +108,12 @@ if __name__ == "__main__":
                 wb.get(chromePath).open_new_tab(search)
                 print(search)
 
+            #FileNotFoundError causes termination
             except Exception as e:
                 speak("Can't open now, please try again later.")
                 print("Can't open now, please try again later.")
 
-
+        #transcribe a note
         elif "remember that" in query:
             speak("What should I remember")
             data = takecommand()
@@ -161,15 +123,17 @@ if __name__ == "__main__":
             remember.write(data)
             remember.close()
 
+        #read out transcribed notes
         elif "do you remember anything" in query:
             remember = open("data.txt", "r")
             speak("You told me to remember that" + remember.read())
             print("You told me to remember that " + str(remember))
 
+        #perform a screesnhsot
         elif "screenshot" in query:
             screenshot()
             speak("I've taken screenshot, please check it")
 
-
+        #end the program
         elif "offline" in query:
             quit()
