+++ Jarvis/jarvis_actions.py
@@ -1,0 +1,89 @@
+#install dependencies
+import pyttsx3
+import pywin32_system32
+import datetime
+import speech_recognition as sr
+import wikipedia
+import webbrowser as wb
+import os
+import random
+import pyautogui
+
+
+#instantiate python text to speech engine - should choose male/female
+engine = pyttsx3.init()
+
+#######################functions########################
+#text to speach
+def speak(audio):
+    engine.say(audio)
+    engine.runAndWait()
+
+
+#read the time 
+def time():
+    Time = datetime.datetime.now().strftime("%I:%M:%S")
+    speak("the current time is")
+    speak(Time)
+    print("The current time is ", Time)
+
+
+#say the date
+def date():
+    day = int(datetime.datetime.now().day)
+    month = int(datetime.datetime.now().month)
+    year = int(datetime.datetime.now().year)
+    speak("the current date is")
+    speak(day)
+    speak(month)
+    speak(year)
+    print("The current date is " + str(day) + "/" + str(month) + "/" + str(year))
+
+
+#greet the user
+def wishme():
+    print("Welcome back sir!!")
+    speak("Welcome back sir!!")
+
+    hour = datetime.datetime.now().hour
+    if 4 <= hour < 12:
+        speak("Good Morning Sir!!")
+        print("Good Morning Sir!!")
+    elif 12 <= hour < 16:
+        speak("Good Afternoon Sir!!")
+        print("Good Afternoon Sir!!")
+    elif 16 <= hour < 24:
+        speak("Good Evening Sir!!")
+        print("Good Evening Sir!!")
+    else:
+        speak("Good Night Sir, See You Tommorrow")
+
+    speak("Jarvis at your service sir, please tell me how may I help you.")
+    print("Jarvis at your service sir, please tell me how may I help you.")
+
+#take a screenshot
+def screenshot():
+    img = pyautogui.screenshot()
+    img_path = os.path.expanduser("~\\Pictures\\ss.png")
+    img.save(img_path)
+
+
+#receive input
+def takecommand():
+    r = sr.Recognizer()
+    with sr.Microphone() as source:
+        print("Listening...")
+        r.pause_threshold = 1
+        audio = r.listen(source)
+
+    try:
+        print("Recognizing...")
+        query = r.recognize_google(audio, language="en-in")
+        print(query)
+
+    except Exception as e:
+        print(e)
+        speak("Please say that again")
+        return "Try Again"
+
+    return query
